# Speedup tetris environment