--- a/gflownet/envs/tetris.py
+++ b/gflownet/envs/tetris.py
@@ -166,17 +166,17 @@ class Tetris(GFlowNetEnv):
         # to add the new piece
         occupied_rows = board[:, col : col + wp].sum(1).nonzero()
         if len(occupied_rows) == 0:
-            # All rows are unoccupied. Set first occupied row as the row "below" the
+            # All rows are unoccupied. Set highest occupied row as the row "below" the
             # board.
-            first_occupied_row = self.height
+            highest_occupied_row = self.height
         else:
-            first_occupied_row = occupied_rows[0, 0]
+            highest_occupied_row = occupied_rows[0, 0]
 
         # Iteratively attempt to place piece on the board starting from the top.
         # As soon as we reach a row where we can't place the piece because it
         # creates a collision, we can stop the search (since we can't put a piece under
         # this obstacle anyway).
-        starting_row = first_occupied_row - hp
+        starting_row = highest_occupied_row - hp
         lowest_valid_row = None
         for row in range(starting_row, self.height - hp + 1):
             if row == -hp: